day type "learn more" -> dashed border
can you make the box of day type have a dashed border when its not up to date instead of having the learn more text? when you tap on the box, it should do the same as when you tap the "learn more" text. the dashed line shouldnt be exactly on the border itself, i want it slightly inwards. when the day is white day, the dashed line should be dark green (same color of the green day box) and vice versa.