splash animation
i wanna make an animation, a splash screen to replace the loading screen right now (only replace the loading screen when first opening app, not pull-to-refresh) i want it an icon(white) to fly in from the bottom on to a dark green(#043615) background, bounce, and when the loading is done, the whole view (both the dark green bg and the icon) slides outside the screen from the top. This picture shows briefly how it would look like. For now, you can replace the icon with a white circle.